Add dash between site and page title
This used to dynamically flip between showing only the site title or page title, depending on whether the `<h1>` of the main content was still on screen or scrolled away. We might want to bring that back, but for short-term simplicity's sake, we just keep this static and insert a dash.