Expose examples from old docs site
Instead of linking to the old docs site like we used to, expose the same examples here. These are still really out of date and need auditing (see #76), but this at least allows us to keep the content maintained here instead of linking to a site we are deprecating.

Note: this does away with the old example viewer and the example JS apps! I think this is OK: the viewer was another thing to maintain and that nobody was working on; the example JS apps didn't do anything that was special to IPFS or to the distributed web, so they weren't especially relevant here.

License: MIT
Signed-off-by: Rob Brackett <rob@robbrackett.com>