Make it clear what "the community section" means
I arrived at the community section of this documentation by clicking "Community" in the footer of the ipfs.io website. I was trying to find links to all of the social channels (IRC, etc.) for someone who asked me where to discuss IPFS questions. When I got to "See the other links in the community section" I tried scrolling down within the page I was on but felt like I was already in the community section. The nav is so long that there was no hint of community on the portion of it in view, though it actually didn't occur to me to look there. ¯\_(ツ)_/¯ 

Is there a way to link someone to an anchor in the nav or do I just need to describe the process of scrolling in the nav until you find the links? Should they be duplicated in the content of the community section here?